+++ client/src/components/images/tab-1-pic.png
