+++ client/src/components/images/tab-macbook.png
