+++ client/src/components/images/tab-tablet.png
