+++ client/src/components/ladingpage/TabComponent.js
