+++ client/src/components/ladingpage/TabDevices.js
