+++ client/src/components/register/RegisterFooter.js
