+++ client/src/models/Users.js
@@ -1,0 +1,8 @@
+const userSchema = new Schema({
+    name: {
+        type: 'string',
+    },
+    email: {
+        type: 'string',
+    },
+})
