+++ server/models/SecretCode.js
@@ -1,0 +1,15 @@
+const secretCode = new Schema({
+    email: {
+        type: String,
+        required: true,
+    },
+    code: {
+        type: String,
+        required: true,
+    },
+    dateCreated: {
+        type: Date,
+        default: Date.now(),
+        expires: 600,
+    },
+});
