+++ server/models/Users.js
@@ -4,7 +4,6 @@ const userSchema = new mongoose.Schema({
     name: {
         type: String,
         required: true,
-        min: 6,
         max: 255
     },
     email: {
